external/opencore: Fix for fragment clip playback failure when using URI
Calling setdatasource with URI on fragmented clip caused the clip to fail
after first few seconds of playback.

In opencore during seek while read, the offset is typecast to 64bit variable
type. When lseek is called with a negative offset, this translates to a huge
offset value thereby seek failing.

CRs-Fixed: 333628

+++ oscl/oscl/osclio/src/oscl_file_native.cpp
@@ -437,7 +437,7 @@ int32 OsclNativeFile::Seek(TOsclFileOffset offset, Oscl_File::seek_type origin)
                 seekmode = SEEK_END;
 #if OSCL_HAS_LARGE_FILE_SUPPORT
 #ifdef ANDROID
-            TOsclFileOffset seekResult = lseek64(iFileDescriptor, offset, seekmode);
+            TOsclFileOffset seekResult = lseek64(iFileDescriptor, ((uint32)offset), seekmode);
             if (seekResult == -1){
               LOGE("OsclNativeFile::Seek lseek64 failed");
               return -1;
